requirements.txt
simple-salesforce==1.12.6
python-dotenv==1.0.1
requests==2.32.3

# LangGraph & LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9

# OpenAI Realtime API
openai==1.54.3
websockets==12.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
pydantic==2.8.2
dataclasses-json==0.6.3

# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
